[Silver II] Title: A → B, Time: 16 ms, Memory: 4980 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.cc
@@ -2,8 +2,8 @@
 #include<vector>
 #include<queue>
 #include<unordered_map>
-long long a;
-long long b;
+int a;
+int b;
 int ans = -1;
 
 std::unordered_map<long long, int> visited;
